requirements.txt
asgiref==3.8.1
Django==4.2.17
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
